pyproject.toml
[project]
name = "esp3-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.115.0",
    "notebook>=7.4.7",
    "numpy>=2.1.0",
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "pdfminer-six>=20250506",
    "pqdm>=0.2.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.34.0",
]
